Refactor GUI module into dedicated package
