+++ src/saftao/gui/__init__.py
@@ -1,0 +1,7 @@
+"""Aplicação gráfica Resolve SAF-T (AO)."""
+
+from __future__ import annotations
+
+from .app import main
+
+__all__ = ["main"]
